tests/python_tests/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
diffusers==0.33.1
optimum-intel @ git+https://github.com/huggingface/optimum-intel.git@main
numpy<2.0.0; platform_system == "Darwin" and platform_machine == "x86_64"
onnx==1.17.0
pytest
pytest-html
hf_transfer
gguf>=0.10.0

# requirements for specific models
# - hf-tiny-model-private/tiny-random-RoFormerForCausalLM
rjieba
# - baichuan-inc/Baichuan2-7B-Chat
bitsandbytes
# - nomic-ai/gpt4all-falcon
# - Qwen/Qwen-7B
# - Qwen/Qwen-7B-Chat
# - mosaicml/mpt-7b
# - internlm/internlm2-7b
einops
# - Qwen/Qwen-7B
# - Qwen/Qwen-7B-Chat
transformers_stream_generator
# - openbmb/MiniCPM-V-2
torchvision
# - openbmb/MiniCPM-V-2
timm
# - Qwen/Qwen-7B
# - Qwen/Qwen-7B-Chat
# - Salesforce/xgen-7b-8k-base
tiktoken
# - microsoft/biogpt
sacremoses
# - openai/whisper-base
librosa
soundfile
datasets
rouge
